refactor styles so that they are broken out of vue components, and in to own files
