+++ packages/vuepress-theme-odyssey/components/odo-card.vue
@@ -13,22 +13,8 @@
       <slot name="footer" />
     </footer>
   </div>
-
 </template>
-<style lang="scss">
-@import '@okta/odyssey';
 
-.odo-card {
-  padding: $spacing-m $spacing-l;
-  border-radius: 0 32px;
-  background: cv('gray', '000');
-
-  &.is-odo-card-plain {
-    background: transparent;
-  }
-}
-
-</style>
 <script>
 export default {
   name: 'odo-card',
