+++ packages/vuepress-theme-odyssey/components/odo-sidebar.vue
@@ -2,73 +2,36 @@
   <header class="odo-sidebar">
     <div class="odo-sidebar--content">
       <div class="odo-sidebar--header">
-        <slot name="header"></slot>
+        <odo-link class="odo-site--title" href="/">{{$site.title}}</odo-link>
+        <fieldset class="ods-fieldset" v-if="$site.themeConfig.flags.hasSearch">
+          <div class="ods-fieldset-flex">
+            <input class="ods-text-input" type="text" name="search" id="search">
+            <label class="ods-label u-visually-hidden" for="name">Name</label>
+          </div>
+        </fieldset>
       </div>
       <div :class='{
         "odo-sidebar--main": true,
-        "is-overflowing": isOverflowing,
+        "is-odo-sidebar-overflowing": isOverflowing,
       }'>
-        <slot name="main"></slot>
+        <odo-nav type="primary" :nav="Nav.primary" />
       </div>
       <div class="odo-sidebar--footer">
-       <slot name="footer"></slot>
+        <odo-nav type="secondary" :nav="Nav.secondary" />
       </div>
     </div>
   </header>
 </template>
 
-<style lang="scss">
-@import '@okta/odyssey';
-
-.odo-sidebar {
-  padding: $spacing-l $spacing-m 0 $spacing-l;
-
-  @media screen and (max-width: 800px) {
-    display: none;
-  }
-
-  .odo-sidebar--content {
-    display: grid;
-    position: sticky;
-    top: $spacing-l;
-    grid-row-gap: $spacing-s;
-    grid-template-areas:
-      'header'
-      'main'
-      'footer';
-    grid-template-rows: auto 1fr auto;
-    height: 100vh;
-    max-height: calc(100vh - #{$spacing-l});
-  }
-
-  .odo-sidebar--header {
-    grid-area: header;
-    padding: 0  $spacing-xs 0;
-  }
-
-  .odo-sidebar--main {
-    grid-area: main;
-    padding: $spacing-xs  $spacing-xs 0;
-    overflow: scroll;
-
-    &.is-overflowing {
-      border-bottom: 1px solid $border-color-display;
-    }
-  }
-
-  .odo-sidebar--footer {
-    grid-area: footer;
-    padding: $spacing-xs  $spacing-xs $spacing-m;
-  }
-}
-</style>
-
 <script>
+import { resolveNav } from '../utils'
+
 export default {
   name: 'odo-sidebar',
   props: ['nav'],
   components: {
     'odo-link': () => import('./odo-link.vue'),
+    'odo-nav': () => import('../components/odo-nav.vue'),
   },
   data: () => ({
     isOverflowing: null,
@@ -92,10 +55,19 @@ export default {
             });
         }, { threshold: 1 });
 
-        observer.observe(el.querySelectorAll('.ods-nav-vertical')[0]) ;
-      }, 500);
-
+        observer.observe(el.querySelectorAll('.odo-nav')[0]) ;
+      }, 100);
     }
+  },
+  computed: {
+    Nav () {
+      return resolveNav(
+        this.$page,
+        this.$page.regularPath,
+        this.$site,
+        this.$localePath
+      )
+    },
   }
 }
 </script>
