+++ packages/vuepress-theme-odyssey/enhanceApp.js
@@ -3,4 +3,5 @@ export default ({
   options,
   router,
   siteData
-}) => {}
+}) => {
+}
