+++ packages/vuepress-theme-odyssey/layouts/Layout.vue
@@ -1,111 +1,28 @@
 <template>
   <div class="theme-odyssey">
     <a class="odo-skip-content" href="#main">Skip to main content</a>
-
-    <odo-sidebar>
-      <template v-slot:header>
-        <a class="odo-site--title" href="/">{{$site.title}}</a>
-        <fieldset class="ods-fieldset" v-if="$site.themeConfig.flags.hasSearch">
-          <div class="ods-fieldset-flex">
-            <input class="ods-text-input" type="text" name="search" id="search">
-            <label class="ods-label u-visually-hidden" for="name">Name</label>
-          </div>
-        </fieldset>
-      </template>
-      <template v-slot:main>
-        <odo-nav-vertical type="primary" :nav="Nav.primary" />
-      </template>
-      <template v-slot:footer>
-        <odo-nav-vertical type="secondary" :nav="Nav.secondary" />
-      </template>
-    </odo-sidebar>
-
+    <odo-sidebar />
     <main class="odo-main" id="main">
-
       <odo-template-home v-if="$page.frontmatter.template === 'home'" />
       <odo-template-component v-else-if="$page.frontmatter.template === 'component'" />
       <odo-template-index v-else-if="$page.frontmatter.template === 'index'" />
       <odo-template-plain v-else-if="$page.frontmatter.template === 'plain'" />
-      <Content v-else />
-          
+      <Content v-else />   
     </main>
   </div>
 </template>
 
-<style lang="scss">
-@import '@okta/odyssey';
-
-.odo-skip-content {
-  position: absolute;
-  top: auto;
-  left: -10000px;
-  width: 1px;
-  height: 1px;
-  overflow: hidden;
-
-  &:focus {
-    left: 0;
-    width: auto;
-    height: auto;
-  }
-}
-
-.odo-site--title,
-.odo-site--title:visited {
-  display: block;
-  padding: $spacing-xs;
-  color: $text-heading;
-  font-size: $size-title-4;
-  font-weight: 600;
-  text-decoration: none;
-}
-
-.odo-main--content {
-  max-width: calc(960px + #{$spacing-l});
-  margin: $spacing-xl auto;
-  padding: 0 $spacing-m;
-}
-
-</style>
-
 <script>
-export default {
-  name: 'FigureAnatomy',
-  props: {
-    img: {
-      type: String,
-      required: true
-    }
-  }
-}
-</script>
-
-<script>
-import { resolveNav } from '../utils'
+import '../styles/index.scss';
 
 export default {
   components: {    
     'odo-sidebar': () => import('../components/odo-sidebar.vue'),
     'odo-link': () => import('../components/odo-link.vue'),
-    'odo-nav-vertical': () => import('../components/odo-nav-vertical.vue'),
     'odo-template-home': () => import('../templates/odo-template-home.vue'),
     'odo-template-index': () => import('../templates/odo-template-index.vue'),
     'odo-template-plain': () => import('../templates/odo-template-plain.vue'),
     'odo-template-component': () => import('../templates/odo-template-component.vue'),
-  },
-  computed: {
-    Nav () {
-      return resolveNav(
-        this.$page,
-        this.$page.regularPath,
-        this.$site,
-        this.$localePath
-      )
-    },
   }
 }
 </script>
-
-<style lang="scss">
-@import '../styles/index.scss';
-</style>
