+++ packages/vuepress-theme-odyssey/package.json
@@ -4,7 +4,7 @@
   "description": "VuePress theme for Odyssey Documentation",
   "author": "Okta, Inc.",
   "license": "Apache-2.0",
-  "main": "layouts/Layout.vue",
+  "main": "index.js",
   "devDependencies": {
     "@okta/odyssey": "^0.5.0",
     "@okta/odyssey-icons": "^0.5.0",
@@ -13,11 +13,7 @@
     "node-sass": "4.13.1",
     "prismjs": "^1.14.0",
     "sass-loader": "8.0.2",
-    "vue-prism-component": "1.2.0",
-    "vuepress-plugin-clean-urls": "^1.1.1",
-    "vuepress-plugin-container": "^2.1.3"
-  },
-  "dependencies": {
-    "vue-fragment": "^1.5.1"
+    "vue-fragment": "^1.5.1",
+    "vuepress-plugin-clean-urls": "^1.1.1"
   }
 }
