+++ packages/vuepress-theme-odyssey/styles/components/_odo-card.scss
@@ -1,0 +1,9 @@
+.odo-card {
+  padding: $spacing-m $spacing-l;
+  border-radius: 0 32px;
+  background: cv('gray', '000');
+
+  &.is-odo-card-plain {
+    background: transparent;
+  }
+}
