+++ packages/vuepress-theme-odyssey/styles/components/_odo-link.scss
@@ -1,0 +1,2 @@
+.odo-link {
+}
