+++ packages/vuepress-theme-odyssey/styles/components/_odo-sidebar.scss
@@ -1,0 +1,41 @@
+.odo-sidebar {
+  padding: $spacing-l $spacing-m 0 $spacing-l;
+
+  @media screen and (max-width: 800px) {
+    display: none;
+  }
+
+  .odo-sidebar--content {
+    display: grid;
+    position: sticky;
+    top: $spacing-l;
+    grid-row-gap: $spacing-s;
+    grid-template-areas:
+      'header'
+      'main'
+      'footer';
+    grid-template-rows: auto 1fr auto;
+    height: 100vh;
+    max-height: calc(100vh - #{$spacing-l});
+  }
+
+  .odo-sidebar--header {
+    grid-area: header;
+    padding: 0  $spacing-xs 0;
+  }
+
+  .odo-sidebar--main {
+    grid-area: main;
+    padding: $spacing-xs  $spacing-xs 0;
+    overflow: scroll;
+
+    &.is-odo-sidebar-overflowing {
+      border-bottom: 1px solid $border-color-display;
+    }
+  }
+
+  .odo-sidebar--footer {
+    grid-area: footer;
+    padding: $spacing-xs  $spacing-xs $spacing-m;
+  }
+}
